Redestructured the CheckNear() function to no longer call every connected tile when activated.
