+++ Assets/GridSystem/Scripts/WallScript.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f6b0550c3820a0b418a4c7a06666f413
+guid: 7498bb6c5bc61a74d9ba4cd871cc9f68
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
