+++ Assets/GridSystem/Tiles/Wall Variant.prefab
@@ -209,9 +209,12 @@ MonoBehaviour:
   m_GameObject: {fileID: 3882245125731334366}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: f6b0550c3820a0b418a4c7a06666f413, type: 3}
+  m_Script: {fileID: 11500000, guid: 7498bb6c5bc61a74d9ba4cd871cc9f68, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  BlockName: 
+  NSEW: 
+  Updating: 0
 --- !u!4 &4515793159389012580 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: -8679921383154817045, guid: 52194195ba157ec4388fe1900f80c4b0,
